# server/swagger.yaml
openapi: 3.1.0
info:
  title: Library API
  description: API quản lý thư viện sách trực tuyến, hỗ trợ Chủ sách và Người đọc
    theo chuẩn OpenAPI
  version: 1.0.0
servers:
- url: http://localhost:3000
  description: Development Server
paths:
...
  /api/v1/collections:
    get:
      tags:
      - Collections
      summary: Get Collections
      description: Tìm kiếm và lấy danh sách bộ sưu tập
      operationId: get_collections_api_v1_collections_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Trang hiện hành (mặc định là 1)
          default: 1
          title: Page
        description: Trang hiện hành (mặc định là 1)
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Kích thước trang (mặc định là 20)
          default: 20
          title: Size
        description: Kích thước trang (mặc định là 20)
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Từ khóa tìm kiếm chung (tùy chọn)
          title: Q
        description: Từ khóa tìm kiếm chung (tùy chọn)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_CollectionResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Collections
      summary: Create Collection
      description: Tạo bộ sưu tập sách mới
      operationId: create_collection_api_v1_collections_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/collections/{id}:
    get:
      tags:
      - Collections
      summary: Get Collection
      description: Xem chi tiết bộ sưu tập (bao gồm các sách bên trong)
      operationId: get_collection_api_v1_collections__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Collections
      summary: Update Collection
      description: Sửa thông tin bộ sưu tập (yêu cầu là chủ bộ sưu tập)
      operationId: update_collection_api_v1_collections__id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Collections
      summary: Delete Collection
      description: Xóa bộ sưu tập (yêu cầu là chủ bộ sưu tập)
      operationId: delete_collection_api_v1_collections__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v2/collections:
    get:
      tags:
      - Collections
      summary: Get Collections
      description: Tìm kiếm và lấy danh sách bộ sưu tập
      operationId: get_collections_api_v2_collections_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Trang hiện hành (mặc định là 1)
          default: 1
          title: Page
        description: Trang hiện hành (mặc định là 1)
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Kích thước trang (mặc định là 20)
          default: 20
          title: Size
        description: Kích thước trang (mặc định là 20)
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Từ khóa tìm kiếm chung (tùy chọn)
          title: Q
        description: Từ khóa tìm kiếm chung (tùy chọn)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_CollectionResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Collections
      summary: Create Collection
      description: Tạo bộ sưu tập sách mới
      operationId: create_collection_api_v2_collections_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/collections/{id}:
    get:
      tags:
      - Collections
      summary: Get Collection
      description: Xem chi tiết bộ sưu tập (bao gồm các sách bên trong)
      operationId: get_collection_api_v2_collections__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Collections
      summary: Update Collection
      description: Sửa thông tin bộ sưu tập (yêu cầu là chủ bộ sưu tập)
      operationId: update_collection_api_v2_collections__id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Collections
      summary: Delete Collection
      description: Xóa bộ sưu tập (yêu cầu là chủ bộ sưu tập)
      operationId: delete_collection_api_v2_collections__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BookResponse:
      properties:
        title:
          type: string
          title: Title
        author:
          type: string
          title: Author
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        quantity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Quantity
        id:
          type: integer
          title: Id
        owner_id:
          type: integer
          title: Owner Id
        collection_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Collection Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - title
      - author
      - id
      - owner_id
      - collection_id
      - created_at
      title: BookResponse
...
    CollectionCreate:
      properties:
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - title
      title: CollectionCreate
    CollectionDetailResponse:
      properties:
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        id:
          type: integer
          title: Id
        owner_id:
          type: integer
          title: Owner Id
        created_at:
          type: string
          format: date-time
          title: Created At
        books:
          items:
            $ref: '#/components/schemas/BookResponse'
          type: array
          title: Books
          default: []
      type: object
      required:
      - title
      - id
      - owner_id
      - created_at
      title: CollectionDetailResponse
    CollectionResponse:
      properties:
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        id:
          type: integer
          title: Id
        owner_id:
          type: integer
          title: Owner Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - title
      - id
      - owner_id
      - created_at
      title: CollectionResponse
    CollectionUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      title: CollectionUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PageResponse_CollectionResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CollectionResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        total_pages:
          type: integer
          title: Total Pages
      type: object
      required:
      - items
      - total
      - page
      - size
      - total_pages
      title: PageResponse[CollectionResponse]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login